+++ backend/src/hatchling/utils/context.py
@@ -11,11 +11,11 @@ class ContextFormatter(ABC):
     @abstractmethod
     def get_formatters(self):
         """
-        A mapping of supported field names to their respective formatting functions. Each function
+        This returns a mapping of supported field names to their respective formatting functions. Each function
         accepts 2 arguments:
 
-        - the value that was passed to the format call, defaulting to `None`
-        - the modifier, defaulting to an empty string
+        - the `value` that was passed to the format call, defaulting to `None`
+        - the modifier `data`, defaulting to an empty string
         """
 
     @classmethod
@@ -40,7 +40,6 @@ class DefaultContextFormatter(ContextFormatter):
         return {
             '/': self.__format_directory_separator,
             ';': self.__format_path_separator,
-            'args': self.__format_args,
             'env': self.__format_env,
             'home': self.__format_home,
             'root': self.__format_root,
@@ -59,19 +58,16 @@ class DefaultContextFormatter(ContextFormatter):
         return self.format_path(os.path.expanduser('~'), data)
 
     def __format_env(self, value, data):
+        if not data:
+            raise ValueError('The `env` context formatting field requires a modifier')
+
         env_var, separator, default = data.partition(':')
         if env_var in os.environ:
             return os.environ[env_var]
         elif not separator:
-            raise ValueError(f'Environment variable without default must be set: {env_var}')
+            raise ValueError(f'Nonexistent environment variable must set a default: {env_var}')
         else:
             return default
-
-    def __format_args(self, value, data):
-        if value is not None:
-            return value
-        else:
-            return data or ''
 
 
 class Context: